api-gateway-spec-reformatter/requirements.txt
python-dotenv>=0.21.0
pyyaml>=6.0